🎨 Move ComponentManager
