+++ bluemira/base/reactor.py
@@ -27,7 +27,6 @@ from warnings import warn
 
 from rich.progress import track
 
-from bluemira.base.builder import ComponentManager
 from bluemira.base.components import Component
 from bluemira.base.error import ComponentError
 from bluemira.base.look_and_feel import bluemira_print
@@ -149,6 +148,73 @@ class BaseManager(abc.ABC):
                 self._filter_tree(comp, dims_to_show),
                 show=i == len(dims_to_show) - 1,
             )
+
+
+class ComponentManager(BaseManager):
+    """
+    A wrapper around a component tree.
+
+    The purpose of the classes deriving from this is to abstract away
+    the structure of the component tree and provide access to a set of
+    its features. This way a reactor build procedure can be completely
+    agnostic of the structure of component trees, relying instead on
+    a set of methods implemented on concrete `ComponentManager`
+    instances.
+
+    This class can also be used to hold 'construction geometry' that may
+    not be part of the component tree, but was useful in construction
+    of the tree, and could be subsequently useful (e.g., an equilibrium
+    can be solved to get a plasma shape, the equilibrium is not
+    derivable from the plasma component tree, but can be useful in
+    other stages of a reactor build procedure).
+
+    Parameters
+    ----------
+    component_tree: Component
+        The component tree this manager should wrap.
+    """
+
+    def __init__(self, component_tree: Component) -> None:
+        self._component = component_tree
+
+    def component(self) -> Component:
+        """
+        Return the component tree wrapped by this manager.
+        """
+        return self._component
+
+    def show_cad(
+        self,
+        *dims: str,
+        **kwargs,
+    ):
+        """
+        Show the CAD build of the component.
+
+        Parameters
+        ----------
+        *dims
+            The dimension of the reactor to show, typically one of
+            'xz', 'xy', or 'xyz'. (default: 'xyz')
+        """
+        ComponentDisplayer().show_cad(
+            self._filter_tree(
+                self.component(), self._validate_cad_dims(*dims, **kwargs)
+            ),
+            **kwargs,
+        )
+
+    def plot(self, *dims: str):
+        """
+        Plot the component.
+
+        Parameters
+        ----------
+        *dims:
+            The dimension(s) of the reactor to show, 'xz' and/or 'xy'.
+            (default: 'xz')
+        """
+        self._plot_dims(self.component(), self._validate_plot_dims(*dims))
 
 
 class Reactor(BaseManager):
